job-admission-predictor-main/backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
pymongo==4.6.0
PyJWT==2.8.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.0.2
werkzeug==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
scipy==1.14.1
joblib==1.4.2
PyPDF2==3.0.1
python-docx==0.8.11
pdfplumber==0.10.4
